Extract resident continuity module
