+++ v2/orchestrator-rust/src/residents/mod.rs
@@ -1,0 +1,3 @@
+mod continuity;
+
+pub(super) use continuity::*;
